incorporation of docker, added my fastAPI to main.py file, and removed theserver.py file
