+++ Dockerfile
@@ -1,0 +1,17 @@
+# Use the official Python image
+FROM python:3.10
+
+# Set the working directory
+WORKDIR /app
+
+# Copy the project files into the container
+COPY . /app
+
+# Install dependencies
+RUN pip install -r requirements.txt
+
+# Expose the port the FastAPI app will run on
+EXPOSE 8000
+
+# Run the FastAPI server
+CMD ["uvicorn", "app.main:app", "--host", "0.0.0.0", "--port", "8000"]
