+++ docker-compose.yml
@@ -1,0 +1,9 @@
+version: "3.9"
+services:
+  calculator:
+    build: .
+    ports:
+      - "8000:8000"
+    volumes:
+      - .:/app
+    restart: always
